Update 'rata_partisipasi' cast type from float to integer in Laporan model

+++ app/Models/Laporan.php
@@ -36,7 +36,7 @@ final class Laporan extends Model
         'alpa_count' => 'integer',
         'total_kehadiran' => 'integer',
         'rata_nilai' => 'float',
-        'rata_partisipasi' => 'float',
+        'rata_partisipasi' => 'integer',
     ];
 
     /**
